+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [1.7.2] - 2023-12-07
+
 ### Fixed
 
 - Correctly report missing `\bibstyle` command when BibTeX fails.